backend/requirements.txt
alembic
bcrypt
email_validator
fastapi
httpx
Jinja2
passlib
pillow
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose
python-multipart
SQLAlchemy
uvicorn
web3
websockets
qrcode
pytest
pytest-asyncio
fastapi-cache2>=0.2.0
web3>=6.0.0
tenacity>=8.2.0
asyncpg